+++ tests/Http/Middleware/PreventRequestForgeryTest.php
@@ -6,9 +6,9 @@ use Illuminate\Contracts\Encryption\Encrypter;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\Session\Session;
 use Illuminate\Foundation\Http\Middleware\PreventRequestForgery;
+use Illuminate\Http\Exceptions\OriginMismatchException;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
-use Illuminate\Session\OriginMismatchException;
 use Illuminate\Session\TokenMismatchException;
 use Mockery as m;
 use PHPUnit\Framework\TestCase;
